fix ruff check (#162)
## Summary

close #137 
refactor DefaultEmbeddingFunction.
Split it into two classes: OnnxEmbeddingFunction and
DefaultEmbeddingFunction.
DefaultEmbeddingFunction worked based on OnnxEmbeddingFunction.
OnnxEmbeddingFunction can execute model based on ONNX.
