+++ src/pyseekdb/utils/embedding_functions/__init__.py
@@ -13,6 +13,7 @@ from .litellm_base_embedding_function import LiteLLMBaseEmbeddingFunction
 from .mistral_embedding_function import MistralEmbeddingFunction
 from .morph_embedding_function import MorphEmbeddingFunction
 from .ollama_embedding_function import OllamaEmbeddingFunction
+from .onnx_embedding_function import OnnxEmbeddingFunction
 from .openai_base_embedding_function import OpenAIBaseEmbeddingFunction
 from .openai_embedding_function import OpenAIEmbeddingFunction
 from .qwen_embedding_function import QwenEmbeddingFunction
@@ -32,6 +33,7 @@ __all__ = [
     "MistralEmbeddingFunction",
     "MorphEmbeddingFunction",
     "OllamaEmbeddingFunction",
+    "OnnxEmbeddingFunction",
     "OpenAIBaseEmbeddingFunction",
     "OpenAIEmbeddingFunction",
     "QwenEmbeddingFunction",
